backend/requirements.txt
-i https://pypi.org/simple
alembic==1.14.1; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
click==8.1.8; python_version >= '3.7'
flask==3.1.0; python_version >= '3.9'
flask-cors==5.0.0
flask-marshmallow==1.3.0; python_version >= '3.9'
flask-migrate==4.1.0; python_version >= '3.6'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.5; python_version >= '3.7'
mako==1.3.9; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.26.1; python_version >= '3.9'
marshmallow-sqlalchemy==1.4.1; python_version >= '3.9'
packaging==24.2; python_version >= '3.8'
psycopg2==2.9.10; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
sqlalchemy==2.0.38; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
werkzeug==3.1.3; python_version >= '3.9'
